requirements.txt
anyascii==0.3.3
asgiref==3.11.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
defusedxml==0.7.1
Django==5.2.9
django-extensions==4.1
django-filter==25.2
django-modelcluster==6.4.1
django-permissionedforms==0.1
django-stubs-ext==5.2.8
django-taggit==6.1.0
django-tasks==0.9.0
django-treebeard==4.8.0
djangorestframework==3.16.1
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.11
laces==0.1.2
modelsearch==1.1.1
openpyxl==3.1.5
pillow==12.0.0
pillow_heif==1.1.1
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.4
telepath==0.3.1
typing_extensions==4.15.0
urllib3==2.6.0
wagtail==7.2.1
Willow==1.12.0
psycopg2-binary==2.9.11
wagtail-color-panel==1.7.1
wagtail-newsletter[mailchimp,mrml]==0.2.4
django-cast==0.2.50
django-allauth==65.13.1
wagtail-modeladmin==2.2.0
wagtail-cache==3.0.0
mailchimp-marketing==3.0.80